# Chapter04/openapi.yml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/biharck/hands-on/1.0.0
  - description: The server description
    url: https://localhost:3000/hands-on-store/1.0.0
info:
  description: |
    This is a sample store server. You can find
    out more about Swagger at
    [http://Swagger.io](http://Swagger.io)
  version: "1.0.0"
  title: Swagger store
  termsOfService: "http://Swagger.io/terms/"
  contact:
    email: biharck@gmail.com
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
paths:
...
  "/users/{username}":
    get:
      tags:
        - user
      summary: Get user by user name
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: The name that needs to be fetched. Use user1 for testing.
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
            application/xml:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid username supplied
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: User not found
      security:
        - bearerAuth: []
    patch:
      tags:
        - user
      summary: Updated user
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
        - name: username
          in: path
          description: name that need to be updated
          required: true
          schema:
            type: string
      responses:
        "204":
          description: successful operation
        "400":
          description: Invalid user supplied
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: User not found
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        description: Updated user object
        required: true
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        "204":
          description: successful operation
        "400":
          description: Invalid username supplied
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: User not found
      security:
        - bearerAuth: []
...
components:
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        userStatus:
          type: integer
          format: int32
          description: User Status
      xml:
        name: User
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for documentation purposes